pyproject.toml
[project]
name = "polymarket-copy-bot"
version = "0.1.0"
description = "CLI Polymarket copy trading bot with paper-first execution"
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
  "pydantic>=2.8,<3.0",
  "requests>=2.32,<3.0",
  "python-dotenv>=1.0,<2.0",
  "PyYAML>=6.0,<7.0"
]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q"
testpaths = ["tests"]
